Update baselines and add one new spread test case
The baselines are about the same or improved since the switch to
transforms. The new test case doesn't emit optimal code, since it has an
unnecessary call to `__assign`, but it's not bad and easier than
adding more flags to stop flag propogation inside nested object
literals. (for a single argument, `__assign` does
require a check of the arguments list length.)